pyproject.toml
[tool.poetry]
name = "bhn-hawksense-crawler"
version = "0.1.0"
description = "Cralwer for hawkSense"
authors = ["theDataFox <leejoshijones@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
sqlalchemy = "^1.3.19"
scrapy = "^2.3.0"
pillow = "^7.2"
nltk = "^3.5"
tldextract = "^2.2.3"
validators = "^0.18.1"
psycopg2 = "^2.8.6"
python-dotenv = "^0.14.0"
scrapyd = "^1.2.1"
scrapyd-client = { git = 'https://github.com/scrapy/scrapyd-client.git', branch = 'master' }
scrapy-splash = { git = 'https://github.com/scrapinghub/scrapy-splash.git', branch = 'master' }
scrapyrt = { git = 'https://github.com/theDataFox/scrapyrt.git', branch = 'master' }
scrapy-fake-useragent = "^1.4.4"
alembic = "^1.4.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
